TASK #175 Move Uuid4IdentitySourceFactory
