+++ spec/NullDev/BroadwaySkeleton/Handler/CreateBroadwayModelHandlerSpec.php
@@ -13,7 +13,7 @@ use NullDev\PHPUnitSkeleton\PHPUnitTestGenerator;
 use NullDev\Skeleton\Definition\PHP\Parameter;
 use NullDev\Skeleton\Definition\PHP\Types\ClassType;
 use NullDev\Skeleton\Source\ImprovedClassSource;
-use NullDev\Skeleton\SourceFactory\Uuid4IdentitySourceFactory;
+use NullDev\Skeleton\Uuid\SourceFactory\Uuid4IdentitySourceFactory;
 use PhpSpec\ObjectBehavior;
 
 class CreateBroadwayModelHandlerSpec extends ObjectBehavior
