+++ spec/NullDev/Skeleton/Uuid/SourceFactory/Uuid4IdentitySourceFactorySpec.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-namespace spec\NullDev\Skeleton\SourceFactory;
+namespace spec\NullDev\Skeleton\Uuid\SourceFactory;
 
 use NullDev\Skeleton\Definition\PHP\Methods\ConstructorMethod;
 use NullDev\Skeleton\Definition\PHP\Methods\ToStringMethod;
 use NullDev\Skeleton\Definition\PHP\Types\ClassType;
 use NullDev\Skeleton\Source\ClassSourceFactory;
 use NullDev\Skeleton\Source\ImprovedClassSource;
-use NullDev\Skeleton\SourceFactory\Uuid4IdentitySourceFactory;
 use NullDev\Skeleton\Uuid\Definition\PHP\Methods\UuidCreateMethod;
+use NullDev\Skeleton\Uuid\SourceFactory\Uuid4IdentitySourceFactory;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
