+++ tests/NullDev/Skeleton/Uuid/SourceFactory/sample-output/SomeId.output
